pyproject.toml
[tool.poetry]
name = "slack-trigger-pd"
version = "0.1.0"
description = ""
authors = ["Nick Groenen <nick@groenen.me>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.8"
slack-sdk = "^3.10.1"
aiohttp = "^3.7.4"

[tool.poetry.dev-dependencies]
azure-functions = "^1.7.2"
pycodestyle = "^2.7.0"
black = "^21.8b0"
mypy = "^0.910"
pylint = "^2.10.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.pylint.FORMAT]
max-line-length = 88

[tool.pylint."MESSAGES CONTROL"]
disable = [
	"missing-module-docstring",
	"too-few-public-methods",
]

[tool.black]
line-length = 88
